first imp of workflow cards and their details
