+++ packages/legacy/core/App/utils/helpers.ts
@@ -173,8 +173,8 @@ export function formatTime(
         ? momentTime.format(formatString)
         : // if non-english, don't include comma between year and month
         isNonEnglish
-          ? `${momentTime.format(formatString)} ${momentTime.format('YYYY')}`
-          : `${momentTime.format(formatString)}, ${momentTime.format('YYYY')}`
+        ? `${momentTime.format(formatString)} ${momentTime.format('YYYY')}`
+        : `${momentTime.format(formatString)}, ${momentTime.format('YYYY')}`
     if (includeHour) {
       formattedTime = `${formattedTime}, ${momentTime.format(hoursFormat)}`
     }
@@ -346,7 +346,7 @@ export const schemaIdFromRestrictions = (queries?: AnonCredsProofRequestRestrict
   // the '2' here is the enum of the transaction type which, for schemas, is always 2
   const schemaId = rstrWithSchemaId
     ? rstrWithSchemaId.schema_id ||
-    `${rstrWithSchemaId.issuer_did}:2:${rstrWithSchemaId.schema_name}:${rstrWithSchemaId.schema_version}`
+      `${rstrWithSchemaId.issuer_did}:2:${rstrWithSchemaId.schema_name}:${rstrWithSchemaId.schema_version}`
     : ''
   return schemaId
 }
@@ -404,33 +404,33 @@ const evaluateOperation = (attribute: number, pValue: number, pType: AnonCredsPr
  */
 export const evaluatePredicates =
   (fields: Fields, credId?: string) =>
-    (proofCredentialItems: ProofCredentialItems): Predicate[] => {
-      const predicates = proofCredentialItems.predicates
-      if (!predicates || predicates.length == 0) {
-        return []
-      }
-
-      if ((credId && credId != proofCredentialItems.credId) || !proofCredentialItems.credId) {
-        return []
-      }
-
-      const credentialAttributes = getCredentialInfo(proofCredentialItems.credId, fields).map((ci) => ci.attributes)
-
-      return predicates.map((predicate: Predicate) => {
-        const { pType: pType, pValue: pValue, name: field } = predicate
-        let satisfied = false
-
-        if (field) {
-          const attribute = (credentialAttributes.find((attr) => attr[field] != undefined) ?? {})[field]
-
-          if (attribute && pValue) {
-            satisfied = evaluateOperation(Number(attribute), Number(pValue), pType as AnonCredsPredicateType)
-          }
-        }
-
-        return { ...predicate, satisfied }
-      })
+  (proofCredentialItems: ProofCredentialItems): Predicate[] => {
+    const predicates = proofCredentialItems.predicates
+    if (!predicates || predicates.length == 0) {
+      return []
     }
+
+    if ((credId && credId != proofCredentialItems.credId) || !proofCredentialItems.credId) {
+      return []
+    }
+
+    const credentialAttributes = getCredentialInfo(proofCredentialItems.credId, fields).map((ci) => ci.attributes)
+
+    return predicates.map((predicate: Predicate) => {
+      const { pType: pType, pValue: pValue, name: field } = predicate
+      let satisfied = false
+
+      if (field) {
+        const attribute = (credentialAttributes.find((attr) => attr[field] != undefined) ?? {})[field]
+
+        if (attribute && pValue) {
+          satisfied = evaluateOperation(Number(attribute), Number(pValue), pType as AnonCredsPredicateType)
+        }
+      }
+
+      return { ...predicate, satisfied }
+    })
+  }
 
 const addMissingDisplayAttributes = (attrReq: AnonCredsRequestedAttribute) => {
   const { name, names, restrictions } = attrReq
@@ -705,25 +705,25 @@ export const retrieveCredentialsForProof = async (
         // We should ignore the key, if the value is undefined. For now this is a workaround.
         ...(hasIndy
           ? {
-            indy: {
-              // Setting `filterByNonRevocationRequirements` to `false` returns all
-              // credentials even if they are revokable (and revoked). We need this to
-              // be able to show why a proof cannot be satisfied. Otherwise we can only
-              // show failure.
-              filterByNonRevocationRequirements: true,
-            },
-          }
+              indy: {
+                // Setting `filterByNonRevocationRequirements` to `false` returns all
+                // credentials even if they are revokable (and revoked). We need this to
+                // be able to show why a proof cannot be satisfied. Otherwise we can only
+                // show failure.
+                filterByNonRevocationRequirements: true,
+              },
+            }
           : {}),
         ...(hasAnonCreds
           ? {
-            anoncreds: {
-              // Setting `filterByNonRevocationRequirements` to `false` returns all
-              // credentials even if they are revokable (and revoked). We need this to
-              // be able to show why a proof cannot be satisfied. Otherwise we can only
-              // show failure.
-              filterByNonRevocationRequirements: true,
-            },
-          }
+              anoncreds: {
+                // Setting `filterByNonRevocationRequirements` to `false` returns all
+                // credentials even if they are revokable (and revoked). We need this to
+                // be able to show why a proof cannot be satisfied. Otherwise we can only
+                // show failure.
+                filterByNonRevocationRequirements: true,
+              },
+            }
           : {}),
         ...(hasPresentationExchange ? { presentationExchange: {} } : {}),
       },
@@ -735,25 +735,25 @@ export const retrieveCredentialsForProof = async (
         // We should ignore the key, if the value is undefined. For now this is a workaround.
         ...(hasIndy
           ? {
-            indy: {
-              // Setting `filterByNonRevocationRequirements` to `false` returns all
-              // credentials even if they are revokable (and revoked). We need this to
-              // be able to show why a proof cannot be satisfied. Otherwise we can only
-              // show failure.
-              filterByNonRevocationRequirements: false,
-            },
-          }
+              indy: {
+                // Setting `filterByNonRevocationRequirements` to `false` returns all
+                // credentials even if they are revokable (and revoked). We need this to
+                // be able to show why a proof cannot be satisfied. Otherwise we can only
+                // show failure.
+                filterByNonRevocationRequirements: false,
+              },
+            }
           : {}),
         ...(hasAnonCreds
           ? {
-            anoncreds: {
-              // Setting `filterByNonRevocationRequirements` to `false` returns all
-              // credentials even if they are revokable (and revoked). We need this to
-              // be able to show why a proof cannot be satisfied. Otherwise we can only
-              // show failure.
-              filterByNonRevocationRequirements: false,
-            },
-          }
+              anoncreds: {
+                // Setting `filterByNonRevocationRequirements` to `false` returns all
+                // credentials even if they are revokable (and revoked). We need this to
+                // be able to show why a proof cannot be satisfied. Otherwise we can only
+                // show failure.
+                filterByNonRevocationRequirements: false,
+              },
+            }
           : {}),
         ...(hasPresentationExchange ? { presentationExchange: {} } : {}),
       },
@@ -1022,8 +1022,8 @@ export const connectFromScanOrDeepLink = async (
     if (receivedInvitation?.id) {
       // Check if this is a QR code scan for a new connection
       if (!isDeepLink) {
-        navigation.navigate(Screens.DummyScreen, {
-          oobRecordId: receivedInvitation.id
+        navigation.navigate(Screens.Workflows, {
+          oobRecordId: receivedInvitation.id,
         })
       } else {
         // Existing behavior
